# Conflicts:
#	src/main/resources/templates/designerView/DesignerDashBoard.html

+++ src/main/java/main/controller/ManagerController.java
@@ -108,7 +108,9 @@ public class ManagerController {
 		String user = userService.getCurrentUser().getName();
 		List<Notification> myNotes = notificationService.findAllByUser(user);
 		for (Notification n : myNotes) {
-			notificationService.delete(n.getId());
+			if (n.getOrder() == orderService.get(id).getId()) {
+				notificationService.delete(n.getId());
+			}
 		}
 
 		return model;
